refactor: move mergeWords function
